+++ R/data.R
@@ -3,7 +3,7 @@
 #' A dataset containing the urls and other attributes of geoservices 
 #' for Uruguay.
 #' @family data
-#' @format A data frame with 59 rows and 9 variables:
+#' @format A data frame with 59 rows and 10 variables:
 #' \describe{
 #'   \item{capa}{name of the geoservice}
 #'   \item{productor}{name of the institution produced the data}
@@ -14,6 +14,7 @@
 #'   \item{url}{url of the service}
 #'   \item{cod}{name of the variable that contains the cod value of the geometries}
 #'   \item{name}{name of the variable that contains the name of the geometries}
+#'   \item{enc}{name of the encoding of the geoservice table}
 #' }
 "metadata"